% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_CalcStats.R
\name{CalcMargStats}
\alias{CalcMargStats}
\title{Calculate Marginal Summary Stats}
\usage{
CalcMargStats(y0, n0, y1, n1, weights = NULL, alpha = 0.05)
}
\arguments{
\item{y0}{Events per category in arm 0.}

\item{n0}{Subjects per category in arm 0.}

\item{y1}{Events per category in arm 1.}

\item{n1}{Subjects per category in arm 1.}

\item{weights}{Stratum mixing weights.}

\item{alpha}{Type I error.}
}
\value{
List containing two data.frames, 
\itemize{
  \item 'Rates', data.frame of marginal rates.
  \item 'Stats', data.frame of marginal contrasts.
}
}
\description{
Calculate the marginal event rates and summary stats starting from stratified
event counts.
}
