% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Main.R
\name{CompMargRates}
\alias{CompMargRates}
\title{Compare Marginal Event Rates.}
\usage{
CompMargRates(
  y0,
  n0,
  y1,
  n1,
  weights = NULL,
  alpha = 0.05,
  boot = FALSE,
  perm = FALSE,
  reps = 2000,
  exclude_double_zero = FALSE
)
}
\arguments{
\item{y0}{Events per category in arm 0.}

\item{n0}{Subjects per category in arm 0.}

\item{y1}{Events per category in arm 1.}

\item{n1}{Subjects per category in arm 1.}

\item{weights}{Stratum mixing weights. If omitted, defaults to the 
stratum proportions.}

\item{alpha}{Type I error.}

\item{boot}{Calculate bootstrap confidence intervals?}

\item{perm}{Perform a permutation-type test of the null?}

\item{reps}{Replicates for bootstrap/permutation.}

\item{exclude_double_zero}{Exclude strata with no events in either arm?}
}
\value{
Object of class `margRates` containing these slots:
\itemize{
  \item `@Rates`, the marginal event rates in each arm.
  \item `@RD`, risk difference analysis.
  \item `@RR`, risk ratio analysis.
  \item `@OR`, odds ratio analysis.
}
}
\description{
Given stratified event count data for two arms, analyze the marginal
event rates via the risk difference, risk ratio, and odds ratio.
}
