% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Class.R
\name{show,margRates-method}
\alias{show,margRates-method}
\title{Show Method}
\usage{
\S4method{show}{margRates}(object)
}
\arguments{
\item{object}{An object of class \code{margRates}.}
}
\description{
Show Method
}
