% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{margOR-class}
\alias{margOR-class}
\title{Compare CICs Object}
\description{
Defines the object class returned by \code{\link{MargOR}}.
}
\section{Slots}{

\describe{
\item{\code{CIs}}{Confidence intervals.}

\item{\code{Rates}}{Marginal rates.}
}}

