% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Rates.R
\name{MargRate}
\alias{MargRate}
\title{Marginal Rates}
\usage{
MargRate(y0, n0, y1, n1)
}
\arguments{
\item{y0}{Events per category in arm 0.}

\item{n0}{Subjects per category in arm 0.}

\item{y1}{Events per category in arm 1.}

\item{n1}{Subjects per category in arm 1.}
}
\value{
Data.frame containing the 'Arm', 'N', and the marginal 'Rate'.
}
\description{
Calculates the marginal event rates in each arm, weighting by the total 
stratum size.
}
