% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Class.R
\docType{class}
\name{margRates-class}
\alias{margRates-class}
\title{Marginal event rates object.}
\description{
Marginal event rates object.
}
\section{Slots}{

\describe{
\item{\code{Rates}}{Marginal rates.}

\item{\code{RD}}{Risk difference.}

\item{\code{RR}}{Risk ratio.}

\item{\code{OR}}{Odds ratio.}
}}

