% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OddsRatio.R
\name{OddsRatio}
\alias{OddsRatio}
\title{Asymptotic Odds Ratio}
\usage{
OddsRatio(p0, n0, p1, n1, alpha = 0.05)
}
\arguments{
\item{p0}{Proportion in arm 0.}

\item{n0}{Sample size in arm 0.}

\item{p1}{Proportion in arm 1.}

\item{n1}{Sample size in arm 1.}

\item{alpha}{Type 1 error rate.}
}
\value{
Data.frame containing:
\itemize{
  \item The odds ratio 'OR'.
  \item The standard error 'SE' of the *log* OR.
  \item The 'Lower' and 'Upper' confidence bound.
  \item The asymptotic 'P' value.
}
}
\description{
Asymptotic Odds Ratio
}
