% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.margOR}
\alias{print.margOR}
\title{Print Method}
\usage{
\method{print}{margOR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{margOR}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{margOR}.
}
