% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Class.R
\name{print.margRates}
\alias{print.margRates}
\title{Print Method}
\usage{
\method{print}{margRates}(x, ...)
}
\arguments{
\item{x}{An object of class \code{margRates}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{margRates}.
}
