% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Statistics.R
\name{RiskRatio}
\alias{RiskRatio}
\title{Asymptotic Risk Difference}
\usage{
RiskRatio(p0, n0, p1, n1, alpha = 0.05)
}
\arguments{
\item{p0}{Marginal proportion in arm 0.}

\item{n0}{Sample size in arm 0.}

\item{p1}{Marginal proportion in arm 1.}

\item{n1}{Sample size in arm 1.}

\item{alpha}{Type 1 error rate.}
}
\value{
Data.frame containing:
\itemize{
  \item The odds ratio 'RR', arm 1 over arm 0.
  \item The standard error 'SE'. 
  \item The 'Lower' and 'Upper' confidence bound.
  \item The asymptotic 'P' value.
}
}
\description{
Asymptotic Risk Difference
}
