% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Boostrap.R
\name{Stats.Boot}
\alias{Stats.Boot}
\title{Bootstrap Summary Statistics.}
\usage{
Stats.Boot(y0, n0, y1, n1, weights = NULL, alpha = 0.05, reps)
}
\arguments{
\item{y0}{Events per category in arm 0.}

\item{n0}{Subjects per category in arm 0.}

\item{y1}{Events per category in arm 1.}

\item{n1}{Subjects per category in arm 1.}

\item{weights}{Stratum mixing weights.}

\item{alpha}{Type 1 error rate.}

\item{reps}{Bootstrap replicates.}
}
\value{
Data.frame containing:
}
\description{
Bootstrap Summary Statistics.
}
