% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Permutation.R
\name{Test.Null}
\alias{Test.Null}
\title{Test the Null Hypothesis via Bootstrap}
\usage{
Test.Null(y0, n0, y1, n1, alpha = 0.05, reps)
}
\arguments{
\item{y0}{Events per category in arm 0.}

\item{n0}{Subjects per category in arm 0.}

\item{y1}{Events per category in arm 1.}

\item{n1}{Subjects per category in arm 1.}

\item{alpha}{Type 1 error rate.}

\item{reps}{Bootstrap replicates.}
}
\value{
Data.frame containing:
}
\description{
Test the Null Hypothesis via Bootstrap
}
